requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==2.1.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
orm-choices==1.0.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
